ADMG/requirements.txt
asttokens==2.4.1
beautifulsoup4==4.12.3
bs4==0.0.2
build==1.0.3
CacheControl==0.13.1
CausalDisco==0.2.0
cdt==0.6.0
certifi==2023.7.22
charset-normalizer==3.3.2
cleo==2.1.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
dulwich==0.21.6
executing==2.0.1
fastjsonschema==2.18.1
filelock==3.13.1
fonttools==4.44.0
fsspec==2023.10.0
gcastle==1.0.3
GPUtil==1.4.0
idna==3.4
igraph==0.11.3
importlib-metadata==6.8.0
installer==0.7.0
ipykernel==6.26.0
ipython==8.17.2
ipywidgets==8.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jupyterlab_widgets==3.0.10
keyring==24.2.0
kiwisolver==1.4.5
llvmlite==0.41.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==2.5
numba==0.58.1
numpy==1.26.4
packaging==23.2
pandas==2.1.3
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-igraph==0.11.3
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
pyzmq==25.1.1
rapidfuzz==3.5.2
requests==2.31.0
requests-toolbelt==1.0.0
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.2
shap==0.43.0
shellingham==1.5.4
six==1.16.0
skrebate==0.62
slicer==0.0.7
soupsieve==2.5
stack-data==0.6.3
statsmodels==0.14.1
sympy==1.12
texttable==1.7.0
threadpoolctl==3.2.0
tomlkit==0.12.2
torch==2.1.2+cu121
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
trove-classifiers==2023.10.18
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.6
wcwidth==0.2.9
widgetsnbextension==4.0.10
zipp==3.17.0
scipy==1.5.2
pandas==1.2.2
ananke-causal==0.1.5
